requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0
pydantic>=2.11.0
anthropic==0.40.0
python-dotenv==1.0.1
duckduckgo-search>=6.3.0